% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_close.R
\name{get_close}
\alias{get_close}
\title{Scrape daily close prices for a list of tickers}
\usage{
get_close(tic_list)
}
\arguments{
\item{tic_list}{a list of tickers to be scraped}
}
\description{
This function scrapes stock closing price data from google using quantmod. The data span 2007-present.
}
\examples{
get_close()

}
\keyword{finance}
\keyword{quantmod}
