% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ret.R
\name{make_ret}
\alias{make_ret}
\title{Make the price data into returns}
\usage{
make_ret(data_df, date_var = "date_id")
}
\arguments{
\item{data_df}{the data frame with stock returns}

\item{date_var}{date variable in quotes. Default to date_id (generated by get_close)}
}
\description{
This function transforms closing price data into stock returns.
}
\examples{
make_ret()
}
\keyword{returns}
