% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/month_ret.R
\name{month_ret}
\alias{month_ret}
\title{make monthly stock return or variance}
\usage{
month_ret(data_df, date_var = "date_id", type, round = FALSE)
}
\arguments{
\item{data_df}{data frame}

\item{date_var}{the name of the date variable in quotes. Default to date_id}

\item{type}{string. The method used for calculating the return or variance}

\item{round}{integer. Option to reduce the degrees of precision. Default to false.}
}
\description{
This function makes monthly stock returns or variances using one of three
  different methods specified by type.
  Type can take on the values: mean, cum, duffee, var
}
\examples{
month_ret()
}
\keyword{returns}
