% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm_download.R
\name{qm_download}
\alias{qm_download}
\title{Scrape Data using quantmod}
\usage{
qm_download(tic)
}
\arguments{
\item{tic}{the ticker of the stock to download}
}
\description{
This function scrapes stock closing price data from google using quantmod. The data span 2007-present.
}
\examples{
qm_download()

}
\keyword{finance}
\keyword{quantmod}
