% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ishares.R
\name{read_ishares}
\alias{read_ishares}
\title{Get detailed ETF data from ishares}
\usage{
read_ishares(ishares_url, etf, save_dir = FALSE)
}
\arguments{
\item{ishares_url}{the url of the file to be downloaded}

\item{etf}{the name of the etf that is downloaded}

\item{save_dir}{directory to save the file, default to FALSE}
}
\description{
This function scrapes an ishares url and selects the relevant variables
  and formats it to be used in R. Option to save the dataframe
}
\examples{
read_ishares()

}
\keyword{finance}
\keyword{import}
\keyword{ishares}
\keyword{scrape}
