% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ret_compare.R
\name{ret_compare}
\alias{ret_compare}
\title{Make different types of returns}
\usage{
ret_compare(data_df)
}
\arguments{
\item{data_df}{data frame with ticker and date variables}
}
\description{
This function makes weekly and monthly returns for a single ticker. It calls the function ret_type.
}
\examples{
ret_compare()
}
\keyword{returns}
