% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ret_type.R
\name{ret_type}
\alias{ret_type}
\title{Create stock returns and variance of a given duration}
\usage{
ret_type(data_df, N.lag, P, l.name)
}
\arguments{
\item{data_df}{data frame}

\item{N.lag}{Number of lags}

\item{P}{the power needed to annualize those lags}

\item{l.name}{the name of the type of return e.g. weekly, monthly}
}
\description{
This function takes a single stock and creates the return based on the number of lags and then returns
  the non-missing remaining data and their variances
}
\examples{
ret_type()
}
\keyword{return}
\keyword{variance}
