% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_NAret.R
\name{rm_NAret}
\alias{rm_NAret}
\title{Find the start of the data or filter stocks that start after a certain date.}
\usage{
rm_NAret(df, start_dt = NA)
}
\arguments{
\item{df}{the dataframe of stock prices to be manipulated}

\item{start_dt}{character yyyy-mm-dd}
}
\description{
This function finds the minimum date for which all stocks in a df have
  return data and removes all dates prior to that date. When a start date
  is specified, all stocks for which the data starts after that date are
  removed.
}
\examples{
rm_NAret(IYE.df, start_dt = "2012-01-01")

}
\keyword{finance}
