% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tic_summ.R
\name{tic_summ}
\alias{tic_summ}
\title{Return the avg ret and sd for individual stocks}
\usage{
tic_summ(data_df, date_var = "date_id")
}
\arguments{
\item{data_df}{the dataframe to be manipulated}

\item{date_var}{the name of the date variable in quotes. Default to date_id}
}
\description{
This function takes a dataframe with stock returns and calculates the mean
  return and the standard deviation of the return for each stock.
}
\examples{
tic_summ()

}
\keyword{finance}
\keyword{mean}
\keyword{sd}
