% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptf_wts.R
\name{ptf_wts}
\alias{ptf_wts}
\title{Calculate the portfolio weights of each stock for the Markowitz efficient ptf}
\usage{
ptf_wts(data_df, date_var, ptf_mu = NA)
}
\arguments{
\item{data_df}{the dataframe to be manipulated}

\item{date_var}{the name of the date variable in quotes}

\item{ptf_mu}{the E ret of the portfolio}
}
\description{
In more detail ... This function
}
\examples{
ptf_weights()

}
\keyword{finance}
