% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_frontier.R
\name{solve_frontier}
\alias{solve_frontier}
\title{calculate the efficient frontier}
\usage{
solve_frontier(data_df, date_var, min_ret = 1.0001, max_ret = 1.01,
  size_out = 100)
}
\arguments{
\item{data_df}{the dataframe to be manipulated}

\item{date_var}{the name of the date variable in quotes}

\item{min_ret}{the minimum return}

\item{max_ret}{the maximum return}

\item{size_out}{the number of obs in the returned df}
}
\description{
In more detail ... This function
}
\examples{
solve_frontier()

}
\keyword{bullet}
\keyword{finance}
\keyword{markowitz}
