% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tic_summ.R
\name{tic_summ}
\alias{tic_summ}
\title{Return the avg ret and variance for individual stocks}
\usage{
tic_summ(df, st.dev = FALSE)
}
\arguments{
\item{df}{the dataframe to be manipulated}

\item{st.dev}{FALSE--returns variance, else, sd.}
}
\description{
This function takes a dataframe with stock returns and calculates the mean
  return and the standard deviation of the return for each stock.
}
\examples{
tic_summ()

}
\keyword{finance}
\keyword{mean}
\keyword{sd}
